ntheory: get rid of the static_casts

+++ src/ntheory.h
@@ -15,9 +15,8 @@ namespace CSymPy {
 inline RCP<Integer> gcd(const Integer &a, const Integer &b)
 {
     mpz_class g;
-    
-    mpz_gcd(g.get_mpz_t(), static_cast<Integer>(a).as_mpz().get_mpz_t(), 
-                                  static_cast<Integer>(b).as_mpz().get_mpz_t());
+
+    mpz_gcd(g.get_mpz_t(), a.as_mpz().get_mpz_t(), b.as_mpz().get_mpz_t());
 
     return rcp(new Integer(g));
 }
@@ -26,23 +25,21 @@ inline RCP<Integer> lcm(const Integer &a, const Integer &b)
 {
     mpz_class c;
 
-    mpz_lcm(c.get_mpz_t(), static_cast<Integer>(a).as_mpz().get_mpz_t(),
-                                 static_cast<Integer>(b).as_mpz().get_mpz_t());
+    mpz_lcm(c.get_mpz_t(), a.as_mpz().get_mpz_t(), b.as_mpz().get_mpz_t());
 
     return rcp(new Integer(c));
 }
 
 inline RCP<Integer> probab_prime_p(const Integer &a, int reps = 25)
 {
-    return rcp(new Integer(mpz_probab_prime_p(
-                         static_cast<Integer>(a).as_mpz().get_mpz_t(), reps)));
+    return rcp(new Integer(mpz_probab_prime_p(a.as_mpz().get_mpz_t(), reps)));
 }
 
 inline RCP<Integer> nextprime(const Integer &a)
 {
     mpz_class c;
 
-    mpz_nextprime(c.get_mpz_t(), static_cast<Integer>(a).as_mpz().get_mpz_t());
+    mpz_nextprime(c.get_mpz_t(), a.as_mpz().get_mpz_t());
 
     return rcp(new Integer(c));
 }
